Use Dependabot to check for outdated action versions used in workflows
Dependabot will periodically check the GitHub Actions workflows and workflow templates of the repository and submit pull
requests to update any that are found to be using an outdated version of an action.

This check has previously been a manual process, and a frequently neglected one. Up to date action versions are
especially important in the workflow templates, since they are the source for the CI systems of many repositories.

Typically, the best approach is to only pin the major version of the action (e.g., `uses: foo/bar@v1`), so the workflow
will automatically use the latest version of the action except when there have been breaking changes. Many actions
provide major version refs for this purpose.

NOTE: Dependabot's PRs will occasionally try to pin to the patch version of the action (e.g., updating `uses: foo/bar@v1`
to `uses: foo/bar@v2.3.4`). When the action author has provided a major version ref, use that instead
(e.g., `uses: foo/bar@v2`). Once the major version has been updated in the workflow, Dependabot should not submit an
update PR again until the next major version bump. So even if the PRs from Dependabot are not always exactly correct,
their value lies in bringing the maintainer's attention to the fact that the action version in use is outdated.
Dependabot will automatically close its PR once the workflow has been updated.

In the case of PRs from Dependabot for updates to the workflow template copies, these serve solely as update
notifications Dependabot has been configured to prefix the PR message with "(DO NOT MERGE)" to make this clear.

I have configured Dependabot to assign me to these pull requests and will take responsibility for reviewing and testing
them.

---
In addition to setting up Dependabot for this repository, I have also provided a template Dependabot configuration file
and documentation to make it easy to setting this up in other repositories where the maintainers have determined it would
be useful.