Add CI workflow to validate configuration files against JSON schema
Some of the assets used by the template workflows have JSON schema that defines the data structure in a machine
readable way. Validating the assets against the schema whenever they are modified ensures that no problems are introduced.
It is especially important to ensure the assets hosted in this repository are correct, since they will be distributed
from here to many other projects.