Use a task for syncing file duplicates
This repository contains intentionally duplicated template workflow copies under
`workflow-templates/dependabot/workflow-template-copies`. These are used for the Dependabot checks for outdated action
versions.

It's essential that these be kept in sync with the source files, and this is enforced by a CI workflow. So the sync
process will need to be run by the developer whenever modifying the template workflows. A Bash script was previously
provided to facilitate syncing, and used by the CI system. However, the Tooling Team standard is to use Task for
development processes that are regularly run locally. So a Task-based approach is most appropriate.