Use a dedicated workflow for each sync type
The approach inherited from the original Bash script-based approach to syncing the repository's file duplicates was to
handle all syncing in a monolithic manner.

As I convert this to a task-based aproach, I have decided that there really isn't enough in common between the types of
files synced, nor their sync method, to justify combining them. Splitting them up allows for path filters on the workflow
triggers and more meaningful results in the event of a failure.