Document template workflow dependency check system
It's important for the template GitHub Actions workflows to use up to date action versions. Dependabot provides just such
a service.

Unfortunately, dependabot's workflow check was designed with the standard `/.github/workflows` workflow location in mind.
For this reason, an odd workaround was necessary to get dependabot checks on the template workflows, which can not be
stored in the standard location. This could be quite confusing to anyone not intimately familiar with the system, so it's
important to provide some documentation. This is linked to from the workflow templates readme, which in turn is linked to
from the root readme.