Attach source comment to task
The idea is for each asset template to provide a link to its source to facilitate syncing improvements to and from the
template. In the case of tasks, most often they will be merged into a single Taskfile.yml. For this reason, placing the
source comment at the top of the template workflow asset Taskfile.yml will lead to the source comment being lost. Placing
the source comment directly above the tasks themselves will increase the chances of the source information being
distributed along with the tasks.