Strip newline from Go build task's LDFLAGS value
The default behavior of the YAML folded block style is to have a final newline. Even though it seems to work fine as-is,
since this variable defines a flag used in a command, it is more appropriate to not have a newline. This is achieved by
adding the "strip chomping indicator".