Use environment variable to define Go distribution dir in release workflows
The Go project release and nightly build workflows contained many hardcoded references to the folder that contains the
build binaries. Even though there will likely not be a need to customize this path for each project, it is set by a
variable in the project-specific taskfile, and so should be in the workflows that rely on that value as well.