Remove Cobra-specific components from versioned website templates
The documentation system of Arduino Lint generates documentation of the command line interface using a custom Go program
and the Cobra command line infrastructure module. This template is intended to provide only a versioned website for any
project containing documentation written in Markdown, whether or not it uses Cobra or even Go.