Add versioned website template documentation
NOTE: the existing versioned websites for tooling projects use outdated versions of mike and mkdocs-material, which
required an approach to creating the version selector menu the author describes as "hacky". That "hacky" approach no
longer works with the modern version of mkdocs-material. Fortunately, this is not a problem since the material theme now
has official support for versioning via mike:
https://squidfunk.github.io/mkdocs-material/setup/setting-up-versioning/
This avoids the need to maintain JavaScript and CSS files.

This does result in a change in the website GUI. The version selector menu has moved from the navigation panel to the
header. Even though I did like the old location, the new one is not so bad either. I like that the version is shown
directly after the project name, making it slightly more clear that it is referring to the project version. Even if the
new UI is slightly worse than the old one, remaining stuck with increasingly outdated dependencies is not an option so we
must move forward.