Do a shallow checkout of repository in nightly Go build workflow
The previous configuration caused a full checkout of the repository. This likely originated as a copy/paste from the Go
release workflow. It is required in that workflow to allow the generation of the changelog from the commit history, but
it is not needed in the nightly, which does not get a changelog.