Remove irrelevant documentation from the versioned website utility script
This information was originally hosted in CONTRIBUTING.md, but it is for maintainers, not contributors so it was not
appropriate for that document. Lacking a better place, I moved it to the script. Now we have a better place in the
dedicated documentation for the versioned website system, which contains this information already. The comments in the
script should be restricted to explanations of its purpose and workings.