Fail Go release template workflows if artifact source files not found
The `actions/upload-artifact` action's default behavior is to only issue a warning if no files are found at the path.
In these usages of the action, the absence of the files which are intended to be uploaded to a workflow artifact
indicates that something has gone horribly wrong. For this reason, the workflow should be configured to fail immediately
if this occurs.