Add template workflow to publish tester builds of Go projects
On every commit to a pull request or the repository:

- Build the project for all supported platforms.
- Upload the builds as workflow artifacts.

This makes it possible for any interested party to participate in beta testing without setting up a build system locally.

This is an unmodified import of the relevant parts of the workflow used by Arduino Lint:
https://github.com/arduino/arduino-lint/blob/5ef264939ff22a4a825df2e44acdd5033dd16e00/.github/workflows/test.yml

There really isn't any good reason for making this part of the test workflow. A dedicated workflow avoids delays in
getting visible results from the unit tests and allows tailoring the workflow trigger's path filter to the specific files
that are relevant to tester builds.