Correct source comment URLs in template workflows
The original plan was for the workflow templates to be hosted in the arduino/.github repository. The workflows that were
the source content already had comments pointing to what was assumed to be their future URL in that repository and I
forgot to update some of them for the new planned location in the arduino/tooling-project-assets repository.