Remove related links from template workflow documentation
At the moment, the related links are far from comprehensive. In their current state, they would do more harm than good by
giving the impression that the ones listed were the only related workflows. At the current state of active development,
it is too difficult to maintain the links. So it's best to just remove them.