Avoid globbing in Markdown link check find command on Windows
For some reason, Task expands the quoted glob when the "Check Markdown" template's `markdown:check-links` task is run on
Windows, and only on Windows. This causes a spurious failure of the task when there is more than one Markdown file in the
repository. On Linux, the glob is passed as a string literal to `find`, which handles it correctly.

I found two workarounds for this annoying behavior:

- Escape the quotes or the glob.
- Use -regex instead of -name.

The `find` command resulting from the first workaround only works on Windows. The `find` command resulting from the
second workaround  works on both Windows and Linux. I feel that having a different command for each operating system
makes the task even more difficult to maintain than it already is, so I opted for the second workaround, and applied it
to the Linux command as well for the sake of consistency (even though the previous command was already working on Linux).