Configure `markdown:check-links` task to check all links on Windows
Task has the odd behavior of treating the markdown-link-check exit status as though the errexit shell option was set,
exiting the task when the first broken link causes a non-zero exit status. But it's preferable to check all links before
exiting with the appropriate exit status to allow the contributor to fix them all in one shot if there are multiple
broken links.

Fortunately, in this case it respects the option value if it is set explicitly in the task.

Task has the opposite behavior on Linux, handling the markdown-link-check exit status as if the errexit shell option as
unset by default, which is the desired result.