Move project-specific environment variable definitions to top of workflow
The goal is for the initial installation and syncs of the template workflows to be as much of a copy/paste operation as
possible. This becomes more difficult when it is necessary to sprinkle project-specific customizations throughout.
Although that is inescapable in certain worflows, when possible, the project-specific parameters should be defined in
environment variables at the top of the workflow.