Remove useless comments re: actions/download-artifact from Go release workflow
Some cryptic comments were added referring to a "v1". These are explained here:
https://github.com/actions/download-artifact#compatibility-between-v1-and-v2

It was definitely important to explain the reason for the need to add the `path` input in the commit/PR that updated the
action from v1 to v2. But this information is not of ongoing value to the reader of the workflow code. We only need to
understand how the version of the action that is currently in use works, and that is fairly clear just from the input
name, with details readily available in the action's documentation. For this reason, the comments only serve to make the
workflow more confusing.