Use consistent approach for installing tools from release assets
I settled on a new approach to installing from release assets when that is the best available source for a tool
dependency (i.e., not available from npm or Poetry) during the development of the "Check Shell Scripts" template workflow.

I think it is beneficial to make the template worflows consistent where possible. Since I think the new approach is a
small improvement over the old one that was previously used in the "Check General Formatting" template workflow, this one
should be updated.