Use website versioning helper script only for generating parameters
Determining the correct versioning parameters for the versioned website system is relatively complex. For this reason,
it's helpful to do this in a Python script, which provides a more capable framework for this code and its testing.

Previously, the system worked like this:

1. Repository event triggers workflow.
2. Workflow executes helper script.
3. Helper script executes task.
4. Task executes mike.

For me, this long chain of scripts, each in a different language, makes the system very difficult to understand and
maintain. Although the helper script is useful for determining the versioning parameters, executing a task is well within
the capabilities of a workflow. The process is run exclusively on GitHub Actions, with no reason to ever run it locally,
to there is no harm to basing significant amounts of it in the workflow. So it's better for the helper script to have a
single job and the workflow to act as the orchestrator of the tools.