Eliminate pointless use of Task for versioned website deploy
The Task task runner allows developers to run common operations for the project via simple commands.

Using these same tasks in the CI/CD system aligns the locally run actions with the automatically run ones.

If an operation should only ever be run from the CI/CD system, putting it in a task only makes the system more convoluted
and difficult to maintain, and risks unknown problems when developers run that task locally. Since the deployment of the
website is handled exclusively by the CD system, using a task for this purpose is inappropriate. The command can simply
be run directly from the workflow.