Make Taskfile installation instructions generic
Most of the template workflows have Taskfiles as assets. Previously, the instructions for installing these assets were
specific to each template. This increases the effort required to maintain the documentation and the likelihood that it
will get out of sync with the content of the assets. After reflection, I decided that this information does not provide a
significant benefit. Generic Taskfile installation instructions should be sufficient for the target reader.