Use POSIX approach to get effective executable path in install script
The installation script checks the path of the effective executable after installing the application. If it is different
from the installation path, it warns the user.

Previously, the external tool `which` was used for this purpose. This is a violation of ShellCheck rule C2230, which
results in an error with the version of ShellCheck installed in the GitHub Actions runner (0.7.0).

The installation script is intended to use POSIX standard tools where possible in order to be as portable as possible.
The recommended replacement `command -v` works equally as well for this particular application, so there is no reason not
to make the change recommended by ShellCheck.

Reference:
https://github.com/koalaman/shellcheck/wiki/SC2230