Adapt installation script test workflow to work with template script
The template installation script will be made non-project specific, to be configured for each project it is applied to.
However, for testing it is necessary to pick an arbitrary application to install.

Although the text replacement approach taken here is crude, it has the advantage of not requiring any modifications be
made to the script solely for the sake of allowing the test. Those modifications would make this important script more
fragile and difficult to maintain. It's better for that fragility and maintenance burden instead be borne by the test
workflow itself.