Replace project-specific configuration value in template installation script
Use of an obvious placeholder value will make it more clear where the template installation script's configuration must
be modified for the individual project it is applied to.

I left the PROJECT_OWNER configuration variable unchanged because it's likely that all repositories where the Tooling
Team uses this script will be under the arduino GitHub organization.