Provide step-by-step template workflow installation instructions
The original documentation format was established in an incarnation of the project where installation usually consisted
only of adding the workflow file to the repository. This time around, the installation is rarely so simple, and often
quite complex. The previous documentation format resulted in the information for the required installation procedures
scattered through the other content of the readme.

A dedicated "Installation" section that contains all the information in a standardized format will make the installation
of the workflows and maintenance of the documentation easier. This is organized into logical subsections:

- Install workflow
- Install assets
- Define dependencies
- Configuration
- Readme badge

If the installer works through these instructions step by step, they should have a fully functional workflow installed
and configured at the end.