Change versioned website template instructions to omit tests
Originally, the tests were integrated into the script, so there was no choice regarding whether they should be installed
along with it. Since then, the tests have been separated. There should not be any need for modifications to the script
code downstream, so it's sufficient to test it here in the upstream repo and the test files would only represent
pointless clutter downstream.