Add template workflow to generate Cobra docs and deploy a versioned MkDocs-based website to GitHub Pages
On every push to the repository's default branch or release branch, generate a command line reference and deploy the
repository's MkDocs-based static website to GitHub Pages.

Documentation content will sometimes apply only to a specific version of the project. For this reason, it's important
for the reader to be able to access the documentation for the specific version of the project they are using.

The documentation system provides access to:

- The tip of the default branch ("dev")
- The latest release ("latest")
- Each minor version series (e.g., "1.2")

The website version is selectable via a menu on the website as well as the URL of each documentation page.

---
This workflow is based on the "Deploy Website" workflow (versioned, MkDocs, Poetry) template.