Link to specific section containing MkDocs dependencies install info
The versioned website template references the documentation from the standard MkDocs website template for all the
information that is the same for both. Since the time it was written, the installation instructions have been organized
into dedicated subsections for each category of installation work. So it is now possible to link directly to the
dependencies installation information.