Report all errors when schema validation occurs
The templates and the repository's CI system use the `ajv-cli` JSON schema validator to check the data format of various
files. The default behavior of `ajv-cli` is to bail out on the first error.

I think it is slightly more convenient to get all the errors at once to allow fixing all of them in one go rather than
having to repeatedly run the validation before discovering all of them. This behavior is provided by adding the
`--all-errors` flag to the `ajv-cli`' commands.