Generate the Cobra docs output folder on demand
The command line reference documentation for Cobra-based Go projects is output to the docs/commands/ folder. Since this
documentation is generated, it is not stored in the repository. Previously, the generation program required this output
folder to already exist. This meant that the otherwise empty folder had to be added to the repository, with a `.gitkeep`
placeholder file since Git does not track empty folders in repositories.

This added extra complexity to the template installation and clutter to the repository. A better approach is for the
documentation generation program to simply create the output folder before generating the content.