Define name for workflow step that runs ShellCheck
The omission of the optional `name` key in the step of the "Check Shell Scripts" workflow that runs ShellCheck caused it
to be identified in the workflow run logs as "Run liskin/gh-problem-matcher-wrap@v1", which is unlikely to mean anything
to the reader.

Use of a descriptive name for this step makes the logs easier to understand, which is especially important for this step
as it is the one that will fail if a script has a ShellCheck rule violation.