Make commit/PR message for "Check ToC" template universally applicable
Previously, the provided commit and PR messages specified that the workflow was used for table of contents in README.md.

Although the stock template workflow does check this file, it is designed to allow it to easily be configured to check
any arbitrary files, and the documentation provides instructions for doing so. The commit and PR messages should be
written so as to be applicable for any installation of the template so that they can simply be copy/pasted without any
modification.