Improve organization of template workflow assets
Most of the template workflows require the installation of configuration files in addition to the workflow itself. These
are referred to as "assets". Since there are often multiple assets with the same filename, they are organized into
folders according to their subject matter.

Some of these assets are either used by multiple template workflows, or have the potential to be used by multiple
workflows as more are added to the repository. It is not be appropriate to store these shared assets in a folder matching
a workflow file name, so a more general folder name is used. This practice was already in effect to some extent, but
sometimes done inconsistently or without enough thought to the general folder names. Since the assets contain comments
providing their source URL, it will be harmful to change their location once the templates start propagating to
downstream projects, so it's worth giving some care to their organization now.