Fix wildcard for YAML filename matching in label sync template
The previous pattern was modeled after the workflow branch filter syntax. In that syntax, the `?` has the same meaning as
in a regular expression, making the preceding character optional. However, the `actions/upload-artifact` action uses
standard globbing syntax, in which the `?` is a placeholder for any single character. The previous incorrect usage of
this syntax caused the intended files to not be uploaded to the artifact.

The action allows providing multiple path patterns in a list, which is now used to accomplish the intended result of
matching both common YAML filename extensions.