Fix glob pattern used to match YAML file extensions in label sync template
The template is intended to support GitHub repository label configuration files that use either of the common file
extensions used for YAML files. The previous brace expansion approach to this caused the command to fail if only one of
the two supported extensions was used by the configuration files:

cat: '.github/label-configuration-files/*.yaml': No such file or directory

This is fixed by using the "or" extended glob syntax.