Refer to Task by correct name in installation steps
The name of the task runner tool is "Task". "Taskfile" is the name of its configuration file. This configuration file is
already in the repository, and if a step could be said to "install" it, that would be the checkout step. So the previous
"Install Taskfile" step name is incorrect and potentially misleading.