Generate tester builds on the modification of any relevant file
Tester builds are intended to facilitate pull request review and beta testing. For the sake of efficiency, the paths
filter should be configured to skip tester build generation on changes that have no effect on the binary, but also so
that they are always generated on any modification to a relevant file. Since the `DistTasks.yml` file is used to build
the application, it is very much relevant.