Add support for multiple modules to "Check Go" workflow
The `check-config` job of the "Check Go" workflow runs a check to see whether there were any missed updates to the
`go.mod` and `go.sum` files. Previously, only a single module was supported. But a project might have multiple modules,
as is the case with any project that uses the Cobra command line reference generation system.

The use of a job matrix will allow the workflow to accommodate any number of modules a project may contain, at any
location.