Make go:lint task compatible with Go 1.16
The update from Go 1.14 to 1.16 broke the task that runs golint. The good news is that the new `go install` command
eliminates the need for the workaround of running the `go get golang.org/x/lint/golint` command from outside the project
path.

The bad news is the `go list` command used to get the path of the golint installation does not work in the "module-aware
mode" that is now the default (and will be the only as of Go 1.17):

missing go.sum entry for module providing package golang.org/x/lint/golint; to add:
        go mod download golang.org/x/lint
task: Command "go list -f {{".Target}}" golang.org/x/lint/golint" failed: exit status 1

In the end, I gave up on making the task work as before. I think it's better to require
the user to install golint and put the installation in the system PATH, displaying a helpful message when this has not
been done.

We plan to evaluate golangci-lint as a replacement for golint, and perhaps this will change the situation, but for now at
least the check is back to functional again.