Add missing install docs for "Check Go" asset
The tasks used by "Check Go" and "Test Go" workflows use the `DEFAULT_GO_PACKAGES` taskfile variable, which provides a
list of all the repository's Go packages. The variable was previously located in the "test-go-task" assets, and so
implicitly covered by the "Test Go" workflow's installation instructions. However, it was not present in the assets for
the "Check Go" workflow.

The solution is to move this shared asset to the general purpose "go-task" assets and add instructions to both
workflows's docs for the installation of that asset.