Remove superfluous checkout step from "Release" workflow
The `create-release` job of the "Release" workflow creates a GitHub release and uploads the built binaries as release
assets. Since the binaries come from the workflow artifact, there is no need for this job to checkout the repository. The
pre-release identification step uses the `GITHUB_REF` environment variable defined by GitHub Actions, which is not
dependent on the runner having a repository checked out.

This means the checkout step serves no purpose. Since it makes the workflow less efficient and more difficult to
maintain, it must be removed.