Run formatting task after documentation generation
The tasks under the `deps` key run in parallel, which is not appropriate for the task that is intended to format the
output from the documentation generation tasks. The generation tasks will typically be able to run concurrently, so the
`go:cli-docs` task can stay in `deps`, but the formatting task must be placed in `cmds`, which will cause it to run after
all tasks in `deps` have finished.