Remove redundant variable from asset taskfile
The installation instructions for all templates that require the `DEFAULT_GO_PACKAGES` taskfile variable point to the
"go-task" asset taskfile as its source. However, there was a redundant copy of this variable in the "check-go-task" asset
taskfile.

Maintaining two copies of this fairly complex asset is difficult, for no benefit in return, so it must be removed.