Make editorconfig-checker exclude regexes explicit
The `Exclude` array in the `.ecrc` configuration file contains the regular expressions of paths that should be excluded
from checking by the editorconfig-checker tool.

Previously, the "template" file contained a list of filenames that should always be ignored. Although it did work, it was
not really correct because the `.` in the filename is actually a regex wildcard, which is not what was intended. Although
a false match was unlikely, this also might mislead users adding project-specific exclusions to the file regarding the
nature of the exclude patterns. Changing to very explicit patterns avoids any chance of false matches and also makes it
clear that these are regexes.